requirements.txt
openai==1.108.0
python-dotenv==1.1.1
Requests==2.32.5
tiktoken==0.11.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
slowapi==0.1.9